+++ bootique-jdbc-test/src/main/java/io/bootique/jdbc/test/runtime/DatabaseChannelFactory.java
@@ -4,6 +4,9 @@ import io.bootique.jdbc.DataSourceFactory;
 import io.bootique.jdbc.test.DatabaseChannel;
 import io.bootique.jdbc.test.DefaultDatabaseChannel;
 
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.util.Collection;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -22,11 +25,11 @@ public class DatabaseChannelFactory {
     }
 
     public DatabaseChannel getChannel() {
-       Collection<String> names = dataSourceFactory.allNames();
-        if(names.size() == 0) {
+        Collection<String> names = dataSourceFactory.allNames();
+        if (names.size() == 0) {
             throw new IllegalStateException("No DataSources configured");
         }
-        if(names.size() > 1) {
+        if (names.size() > 1) {
             throw new IllegalStateException("More than one DataSource is configured. Don't know which one is default");
         }
 
@@ -38,6 +41,20 @@ public class DatabaseChannelFactory {
     }
 
     protected DatabaseChannel createChannel(String dataSourceName) {
-        return new DefaultDatabaseChannel(dataSourceFactory.forName(dataSourceName));
+        DataSource dataSource = dataSourceFactory.forName(dataSourceName);
+
+        String identifierQuote;
+        try (Connection c = dataSource.getConnection()) {
+            identifierQuote = c.getMetaData().getIdentifierQuoteString();
+        } catch (SQLException e) {
+            throw new RuntimeException("Error reading test DB metadata");
+        }
+
+        // if no quotations are supported, the returned value is space, so "trim" should do the trick of converting this
+        // to no-quote
+
+        identifierQuote = identifierQuote.trim();
+
+        return new DefaultDatabaseChannel(dataSource, identifierQuote);
     }
 }