+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #17 Table.selectOne ignores all columns but the first one
 * #20 Test Table.deleteAll(..) should not declare a checked exception
+* #21 Tests: DataChannel to quote SQL identifiers by default
 
 ## 0.12